Tidy up Frontier scenario into smaller files
